+++ src/Bitrix24/Contacts/Transport/Admin/Transport.php
@@ -10,6 +10,7 @@ use B24io\Loyalty\SDK\Bitrix24\Contacts\Transport\DTO\ContactResponse;
 use B24io\Loyalty\SDK\Bitrix24\Contacts\Transport\DTO\FiltrationResultResponse;
 use Fig\Http\Message\RequestMethodInterface;
 use libphonenumber\PhoneNumber;
+use Ramsey\Uuid\UuidInterface;
 
 /**
  * Class Transport
@@ -206,6 +207,51 @@ class Transport extends SDK\Transport\AbstractTransport
         $this->log->debug(
             'b24io.loyalty.sdk.bitrix24.contacts.transport.admin.filterContactsByPhone.finish',
             [
+                'metadata' => SDK\Transport\Formatters\Metadata::toArray($response->getMeta()),
+            ]
+        );
+
+        return $response;
+    }
+
+    /**
+     * @param UuidInterface $cardUuid
+     *
+     * @return ContactResponse
+     * @throws SDK\Exceptions\ApiClientException
+     * @throws SDK\Exceptions\BaseLoyaltyException
+     * @throws SDK\Exceptions\NetworkException
+     * @throws SDK\Exceptions\ObjectInitializationException
+     * @throws SDK\Exceptions\TransportFormatException
+     * @throws SDK\Exceptions\UnknownException
+     */
+    public function getByCardUuid(UuidInterface $cardUuid): ContactResponse
+    {
+        $this->log->debug(
+            'b24io.loyalty.sdk.bitrix24.contacts.transport.admin.getByCardUuid.start',
+            [
+                'cardUuid' => $cardUuid->toString(),
+            ]
+        );
+
+        $requestResult = $this->apiClient->executeApiRequest(
+            sprintf('admin/bitrix24-contacts/get-by-card-uuid/%s/', $cardUuid->toString()),
+            RequestMethodInterface::METHOD_GET
+        );
+
+        $response = new ContactResponse(
+            $this->initMetadata($requestResult['meta']),
+            SDK\Bitrix24\Contacts\DTO\Fabric::initContactFromArray($requestResult['result']['contact']),
+            SDK\Cards\DTO\Fabric::initFromArray($requestResult['result']['card'])
+        );
+
+        $this->log->debug(
+            'b24io.loyalty.sdk.Bitrix24.Contacts.transport.admin.getByCardUuid.finish',
+            [
+                'contact'  => $response->getContact() !== null ? SDK\Bitrix24\Contacts\Formatters\Contact::toArray(
+                    $response->getContact()
+                ) : null,
+                'card'     => $response->getCard() !== null ? SDK\Cards\Formatters\Card::toArray($response->getCard()) : null,
                 'metadata' => SDK\Transport\Formatters\Metadata::toArray($response->getMeta()),
             ]
         );